# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateDespesaFixaRequest:
    properties:
      categoria_id:
        type: string
      descricao:
        maxLength: 255
        minLength: 3
        type: string
      dia_vencimento:
        maximum: 31
        minimum: 1
        type: integer
      fornecedor:
        type: string
      observacoes:
        type: string
      unidade_id:
        type: string
      valor:
        description: Dinheiro sempre string
        type: string
    required:
    - descricao
    - dia_vencimento
    - valor
    type: object
...
  dto.DespesaFixaResponse:
    properties:
      ativo:
        type: boolean
      atualizado_em:
        type: string
      categoria_id:
        type: string
      criado_em:
        type: string
      descricao:
        type: string
      dia_vencimento:
        type: integer
      fornecedor:
        type: string
      id:
        type: string
      observacoes:
        type: string
      tenant_id:
        description: Opcional, pode ser omitido
        type: string
      unidade_id:
        type: string
      valor:
        description: Dinheiro sempre string
        type: string
    type: object
  dto.DespesasFixasListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.DespesaFixaResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.DespesasFixasSummaryResponse:
    properties:
      total:
        type: integer
      total_ativas:
        type: integer
      valor_total:
        description: Dinheiro sempre string
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.GerarContasRequest:
    properties:
      ano:
        description: Se vazio, usa ano atual
        type: integer
      mes:
        description: Se vazio, usa mês atual
        type: integer
    type: object
  dto.GerarContasResponse:
    properties:
      contas_criadas:
        type: integer
      detalhes_erros:
        items:
          type: string
        type: array
      erros:
        type: integer
      tempo_execucao_ms:
        type: integer
      total_despesas:
        type: integer
    type: object
...
  dto.UpdateDespesaFixaRequest:
    properties:
      categoria_id:
        type: string
      descricao:
        maxLength: 255
        minLength: 3
        type: string
      dia_vencimento:
        maximum: 31
        minimum: 1
        type: integer
      fornecedor:
        type: string
      observacoes:
        type: string
      unidade_id:
        type: string
      valor:
        description: Dinheiro sempre string
        type: string
    required:
    - descricao
    - dia_vencimento
    - valor
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@nexo.barber
    name: API Support
    url: https://nexo.barber/support
  description: API completa para gestão de barbearias com módulos de Agendamento,
    Comanda, Financeiro, Estoque e Analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://nexo.barber/terms
  title: NEXO - Barber Analytics API
  version: "2.0"
paths:
...
  /api/v1/financial/fixed-expenses:
    get:
      description: Retorna lista paginada de despesas fixas
      parameters:
      - description: 'Página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (default: 20)'
        in: query
        name: page_size
        type: integer
      - description: Filtrar por status ativo
        in: query
        name: ativo
        type: boolean
      - description: Filtrar por categoria
        in: query
        name: categoria_id
        type: string
      - description: Filtrar por unidade
        in: query
        name: unidade_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DespesasFixasListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar despesas fixas
      tags:
      - Despesas Fixas
    post:
      consumes:
      - application/json
      description: Cria uma nova despesa fixa recorrente
      parameters:
      - description: Dados da despesa fixa
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDespesaFixaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DespesaFixaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar despesa fixa
      tags:
      - Despesas Fixas
  /api/v1/financial/fixed-expenses/{id}:
    delete:
      description: Remove uma despesa fixa
      parameters:
      - description: ID da despesa fixa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Excluir despesa fixa
      tags:
      - Despesas Fixas
    get:
      description: Retorna uma despesa fixa específica
      parameters:
      - description: ID da despesa fixa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DespesaFixaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter despesa fixa por ID
      tags:
      - Despesas Fixas
    put:
      consumes:
      - application/json
      description: Atualiza uma despesa fixa existente
      parameters:
      - description: ID da despesa fixa
        in: path
        name: id
        required: true
        type: string
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDespesaFixaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DespesaFixaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar despesa fixa
      tags:
      - Despesas Fixas
  /api/v1/financial/fixed-expenses/{id}/toggle:
    patch:
      description: Ativa ou desativa uma despesa fixa
      parameters:
      - description: ID da despesa fixa
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DespesaFixaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Alternar status ativo/inativo
      tags:
      - Despesas Fixas
  /api/v1/financial/fixed-expenses/generate:
    post:
      consumes:
      - application/json
      description: Cria contas a pagar para o mês especificado a partir das despesas
        fixas ativas
      parameters:
      - description: Mês/Ano para geração
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GerarContasRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GerarContasResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Gerar contas a pagar a partir de despesas fixas
      tags:
      - Despesas Fixas
  /api/v1/financial/fixed-expenses/summary:
    get:
      description: Retorna totais e soma de valores das despesas fixas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DespesasFixasSummaryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Obter resumo de despesas fixas
      tags:
      - Despesas Fixas
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token JWT (formato: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...